matmul add expand relu -- working

+++ tensor/matmul.cu
@@ -8,10 +8,10 @@ __global__ void matmul(float *a, float *b, float *c, int m, int n, int k)
 {
     int row = blockIdx.y * blockDim.y + threadIdx.y;
     int col = blockIdx.x * blockDim.x + threadIdx.x;
-    int sum = 0;
-    if( col < k && row < m)
+    float sum = 0;
+    if (col < k && row < m)
     {
-        for(int i = 0; i < n; i++)
+        for (int i = 0; i < n; i++)
         {
             sum += a[row * n + i] * b[i * k + col];
         }
